Refactor WEBRTC files to /media for better packaging.
Added method stop() on MediaStreamTrack to stop the current streaming track.
Applied Promise for getUserMedia() as new API suggested for NEW Implementation.
