+++ src/main/java/gwt/material/design/jscore/client/api/media/MediaStream.java
@@ -1,4 +1,4 @@
-package gwt.material.design.jscore.client.api;
+package gwt.material.design.jscore.client.api.media;
 
 /*
  * Copyright 2016 Google Inc.
@@ -16,35 +16,38 @@ package gwt.material.design.jscore.client.api;
  * the License.
  */
 
+import jsinterop.annotations.JsMethod;
 import jsinterop.annotations.JsPackage;
+import jsinterop.annotations.JsProperty;
 import jsinterop.annotations.JsType;
 
 @JsType(isNative = true, namespace = JsPackage.GLOBAL)
-public class URL {
+public class MediaStream {
 
-    public String hash;
-    public String host;
-    public String hostname;
-    public String href;
-    public String origin;
-    public String password;
-    public String pathname;
-    public String port;
-    public String protocol;
-    public String search;
-    public String username;
+    @JsProperty
+    public boolean active;
 
-    public static native String createObjectURL(MediaStream var0);
+    @JsProperty
+    public boolean ended;
 
-    public static native String domainToASCII(String var0);
+    @JsProperty
+    public String id;
 
-    public static native String domainToUnicode(String var0);
+    @JsMethod
+    public native void addTrack(MediaStreamTrack track);
 
-    public static native Object revokeObjectURL(String var0);
+    @JsMethod
+    public native MediaStream clone();
 
-    public URL(String url, String base) {
-    }
+    @JsMethod
+    public native MediaStreamTrack getAudioTracks();
 
-    public URL(String url) {
-    }
+    @JsMethod
+    public native MediaStreamTrack[] getTracks();
+
+    @JsMethod
+    public native MediaStreamTrack getVideoTracks();
+
+    @JsMethod
+    public native void removeTrack(MediaStreamTrack track);
 }
